:zap: (ME-92) AppleRESTAPI 로직 분리, UseCase 의 Rx 전용 함수 분리
